feat: Update views to show user follow status button
This update shows the current state of the user's following to the
authenticated user and invites them to either follow (if not already) or
to unfollow (if following).

Turbo streams and frames are used to make the experience seemless so the
page is not reloaded

TODO: Refactor the code so that there's unique IDs for users who have
different posts that appear on the same page. Currently for such users,
the update doesn't look great yet.