feat: Remove slug option for users and handle it from the model
The slug will now be handled by the model before it is saved. Slugs will
now be the title with in all lowercase with URL-safe base64 string
appended to it.

In the future, I might allow users to specify their preffered unique
slugs. That will be used if present and a fallback to the current
implementation if it fails to be provided by the user.