feat: Fix the flaky issues with the "follow" and "unfollow" buttons
This update finally fixes the issue related with the follow and unfollow
buttons. I had to turn off turbo because I couldn't figure it out in the
beginning. Now, turbo is enable and turbo-stream and turbo-frame have
been used successfully to implement a nice experience which ensures that
there's not a full page reload when either buttons are clicked. Only the
buttons are updated.