feat: Improve the turbo streaming of posts when for fresh starts
I realized in production that Posts were not displayed (streamed) when
starting on a clean slate. This update ensures that the posts are shown
immediately they are created, especially when there are no available
posts at the beginning.

Initially, because there were available posts in development, I didn't
catch this error. Now it's fixed for production as well.