feat: Change from button to anchor tag to avoid false error
Because the button was sending a POST request by default, users saw the
message "Invalid Email or password" even before they had typed anything.
This update changes to a simple link tag so that it defaults to a GET
request.

I could have added the `method: :get` to the already existing code
instead of changing it to `link_to` but I think this is much better.